modificaciones generales

+++ app/Http/Controllers/ReceiptController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\PaymentMethod;
 use App\Receipt;
 use App\Provider;
 use App\Product;
 
 use Carbon\Carbon;
 use App\ReceivedProduct;
+use App\Transaction;
 use Illuminate\Http\Request;
 
 class ReceiptController extends Controller
@@ -72,11 +74,17 @@ class ReceiptController extends Controller
      */
     public function destroy(Receipt $receipt)
     {
-        $receipt->delete();
+        if (!$receipt->finalized_at) {
+            $receipt->delete();
 
-        return redirect()
-            ->route('receipts.index')
-            ->withStatus('Compra eliminada.');
+            return redirect()
+                ->route('receipts.index')
+                ->withStatus('Compra eliminada.');
+        } else {
+            return redirect()
+                ->route('receipts.index')
+                ->withErrors('La compra no puede ser eliminada, ya esta finalizada.');
+        }
     }
 
     /**
@@ -91,7 +99,7 @@ class ReceiptController extends Controller
         $receipt->save();
 
         foreach ($receipt->products as $receivedproduct) {
-            $receivedproduct->product->stock += $receivedproduct->stock;
+            $receivedproduct->product->stock += $receivedproduct->qty;
             $receivedproduct->product->save();
         }
 
@@ -106,11 +114,77 @@ class ReceiptController extends Controller
      */
     public function addproduct(Receipt $receipt)
     {
-        $products = Product::where('type','<>', 2)->get();
-
-        return view('inventory.receipts.addproduct', compact('receipt', 'products'));
+        $products = Product::where('type', '<>', 2)->with('receiveds.receipt')->get();
+        $providerReceipts = Receipt::where('provider_id', $receipt->provider_id)->get();
+        return view('inventory.receipts.addproduct', compact('receipt', 'products', 'providerReceipts'));
     }
 
+    public function addtransaction(Receipt $receipt)
+    {
+
+        $payment_methods = PaymentMethod::all();
+
+        return view('inventory.receipts.addtransaction', compact('receipt', 'payment_methods'));
+    }
+
+
+    public function storetransaction(Request $request, Receipt $receipt, Transaction $transaction)
+    {
+        switch ($request->all()['type']) {
+
+            case 'payment':
+                $request->merge(['title' => 'Pago a proveedor por compra ID: ' . $request->get('receipt_id')]);
+
+                if ($request->get('amount') > 0) {
+                    $request->merge(['amount' => (float) $request->get('amount') * (-1)]);
+                }
+                break;
+        }
+
+        $transaction->create($request->all());
+
+        return redirect()
+            ->route('receipts.show', compact('receipt'))
+            ->withStatus('Transacción registrada.');
+    }
+
+    public function edittransaction(Receipt $receipt, Transaction $transaction)
+    {
+        $payment_methods = PaymentMethod::all();
+
+        return view('inventory.receipts.edittransaction', compact('receipt', 'transaction', 'payment_methods'));
+    }
+
+    public function updatetransaction(Request $request, Receipt $receipt, Transaction $transaction)
+    {
+        switch ($request->get('type')) {
+            case 'income':
+                $request->merge(['title' => 'Ingreso recibido de la venta ID: ' . $request->get('receipt_id')]);
+                break;
+
+            case 'expense':
+                $request->merge(['title' => 'Gasto a cliente por venta ID: ' . $request->get('receipt_id')]);
+
+                if ($request->get('amount') > 0) {
+                    $request->merge(['amount' => (float) $request->get('amount') * (-1)]);
+                }
+                break;
+        }
+        $transaction->update($request->all());
+
+        return redirect()
+            ->route('receipts.show', compact('receipt'))
+            ->withStatus('Transacción modificada.');
+    }
+
+    public function destroytransaction(Receipt $receipt, Transaction $transaction)
+    {
+        $transaction->delete();
+
+        return redirect()
+            ->route('receipts.show', $receipt)
+            ->withStatus('Transacción eliminada.');
+    }
     /**
      * Add product on Receipt.
      *
@@ -120,6 +194,8 @@ class ReceiptController extends Controller
      */
     public function storeproduct(Request $request, Receipt $receipt, ReceivedProduct $receivedproduct)
     {
+        $request->merge(['total_amount' => $request->get('cost') * $request->get('qty')]);
+
         $receivedproduct->create($request->all());
 
         return redirect()
